Ndaj AdminLeads: constants, UI components dhe modals
